# Modify app initialization to speed up loading (PP-893)